+++ test/int/sql/test_datetime.sql
@@ -40,6 +40,18 @@ select '2026-01-13' < '2026-02-01'::datetime;
 -- EXPECTED:
 true
 
+-- TEST: datetime-3.1
+-- SQL:
+select '2026-01-13' between '2026-01-01' and '2026-01-20'::datetime;
+-- EXPECTED:
+true
+
+-- TEST: datetime-3.2
+-- SQL:
+select '2026-01-13' between '2026-01-01'::datetime and '2026-01-20';
+-- EXPECTED:
+true
+
 -- TEST: datetime-3.3
 -- SQL:
 select '2026-01-13' between '2026-01-01'::datetime and '2026-01-20'::datetime;