+++ CHANGELOG.md
@@ -17,6 +17,9 @@ with the `YY.MINOR.MICRO` scheme.
   errors.
 - Fixed that upgrading between patch versions wouldn't run upgrade scripts.
 - Datetime literals should support `yyyy-mm-dd` format, e.g. `select '2026-01-17'::datetime`.
+- Fix type inference for the `a BETWEEN b AND c` expression; now types of `a`, `b` and `c` should be
+  properly unified, meanining that `select '2026-01-13' between '2026-01-01'::datetime and '2026-01-20'`
+  will work as expected.
 
 
 ## [25.5.3] - 2026-01-15